requirements.txt
Flask>=2.3.0
Pillow>=10.0.0
